refactor: JpaRepository 를 infra 로 이동
