+++ src/main/java/kitchenpos/menus/tobe/infra/JpaMenuRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.infra;
 
+import kitchenpos.menus.tobe.domain.Menu;
+import kitchenpos.menus.tobe.domain.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
