+++ src/main/java/kitchenpos/products/tobe/infra/JpaProductRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.tobe.infra;
 
+import kitchenpos.products.tobe.domain.Product;
+import kitchenpos.products.tobe.domain.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
